Revert "feat: "닉네임 초기화 기능 구현""
