--- src/main/java/com/estsoft/guesshangeul/admin/service/AdminService.java
@@ -1,20 +1,0 @@
-package com.estsoft.guesshangeul.admin.service;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import com.estsoft.guesshangeul.user.entity.Users;
-import com.estsoft.guesshangeul.user.repository.UsersRepository;
-
-@Service
-public class AdminService {
-	@Autowired
-	UsersRepository usersRepository;
-
-	public Users resetNickname(Long userId) {
-		Users users = usersRepository.findById(userId).orElseThrow();
-		users.setNickname("hangeul_" + users.getId());
-		return usersRepository.save(users);
-	}
-
-}
