+++ src/main/java/com/estsoft/guesshangeul/entity/QuizPost.java
@@ -1,10 +1,8 @@
-package com.estsoft.guesshangeul.admin.entity;
+package com.estsoft.guesshangeul.entity;
 
 import java.time.LocalDateTime;
 
 import org.hibernate.annotations.ColumnDefault;
-
-import com.estsoft.guesshangeul.user.entity.Users;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -22,21 +20,40 @@ import lombok.Setter;
 @Setter
 @Entity
 @NoArgsConstructor
-@Table(name = "board_manager_apply")
-public class BoardManagerApply {
+@Table(name = "quiz_post")
+public class QuizPost {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "board_manager_apply_id", unique = true, nullable = false)
+	@Column(name = "quiz_post_id", unique = true, nullable = false)
 	private Long id;
 
 	@ManyToOne
 	@JoinColumn(name = "user_id")
-	private Users users;
+	private Users user;
+
+	@ManyToOne
+	@JoinColumn(name = "quiz_board_id")
+	private GeneralBoard generalBoard;
+
+	@Column(name = "quiz_title",nullable = false)
+	private String quizTitle;
+
+	@Column(name = "hint_content",nullable = false)
+	private String hintContent;
 
 	@Column(nullable = false)
-	@ColumnDefault("0")
-	private int status;
+	private String answer;
+
+	@Column(name = "is_hidden", nullable = false)
+	@ColumnDefault("false")
+	private Boolean isHidden;
+
+	@Column(nullable = false)
+	private Long view;
 
 	@Column(name = "created_at")
 	private LocalDateTime createdAt;
+
+	@Column(name = "updated_at")
+	private LocalDateTime updatedAt;
 }
