--- src/main/java/com/estsoft/guesshangeul/user/dto/UsersResponse.java
@@ -1,18 +1,0 @@
-package com.estsoft.guesshangeul.user.dto;
-
-import com.estsoft.guesshangeul.user.entity.Users;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class UsersResponse {
-	Long user_id;
-	String nickname;
-
-	public UsersResponse(Users users) {
-		this.user_id = users.getId();
-		this.nickname = users.getNickname();
-	}
-}
