+++ src/main/resources/application.properties
@@ -1,6 +1,1 @@
 spring.application.name=guess-hangeul
-# JPA configuration
-spring.jpa.show-sql=true
-spring.jpa.properties.hibernate.format_sql=true
-spring.jpa.defer-datasource-initialization=true
-spring.profiles.include=local
